+++ packages/react-craft-ai-operations-history/src/components/operationsHistory.jsx
@@ -73,7 +73,7 @@ function computeUpdatedEstimations({
 
   let updatedFrom;
   let updatedLoadedFrom = Math.min(
-    loadedOperations[loadedCount - 1].timestamp, // Operations are ordered from the latest to the earliest
+    loadedOperations[0].timestamp, // Operations are ordered from the earliest to the latest
     loadedFrom || TIMESTAMP_MAX
   );
   if (from == null) {
@@ -86,7 +86,7 @@ function computeUpdatedEstimations({
 
   let updatedTo;
   let updatedLoadedTo = Math.max(
-    loadedOperations[0].timestamp, // Operations are ordered from the latest to the earliest
+    loadedOperations[loadedCount - 1].timestamp, // Operations are ordered from the earliest to the latest
     loadedTo || updatedLoadedFrom
   );
   if (to == null) {
@@ -98,7 +98,8 @@ function computeUpdatedEstimations({
   }
 
   // Period is the average time between operations
-  const updatedEstimatedPeriod =
+  const updatedEstimatedPeriod = loadedCount === 1 ?
+    loadedCount :
     (updatedLoadedTo - updatedLoadedFrom) / (loadedCount - 1);
 
   // Estimate how many operations there is before and after what is loaded
@@ -125,17 +126,18 @@ function computeUpdatedEstimations({
 }
 
 function computeInitialStateFromProps(props) {
-  const { agentConfiguration, focus, from, initialOperations, to } = props;
+  const { entityConfiguration, focus, from, initialOperations, initialState, to } = props;
 
   const loadedOperations = preprocessOperations(
-    agentConfiguration,
-    initialOperations
+    entityConfiguration,
+    initialOperations,
+    initialState
   );
   const estimations = computeUpdatedEstimations({
     from,
     to,
     loadedOperations,
-    estimatedPeriod: agentConfiguration.time_quantum || 1 // Use a time quantum of 1 as the default period
+    estimatedPeriod: entityConfiguration.time_quantum || 1 // Use a time quantum of 1 as the default period
   });
 
   return {
@@ -259,7 +261,7 @@ class OperationsHistory extends React.Component {
     this._renderRow = this._renderRow.bind(this);
     this._renderPlaceholderRow = this._renderPlaceholderRow.bind(this);
 
-    this._extractedProperties = extractProperties(props.agentConfiguration);
+    this._extractedProperties = extractProperties(props.entityConfiguration);
     this._totalWidth = TIMESTAMP_CELL_WIDTH;
     this._extractedProperties.forEach(
       ({ property }) => (this._totalWidth += computeCellWidth(property.length))
@@ -324,7 +326,7 @@ class OperationsHistory extends React.Component {
             onRequestOperations(requestedFrom, requestedTo, true)
               .then(
                 (result) => {
-                  const { agentConfiguration } = this.props;
+                  const { entityConfiguration } = this.props;
 
                   if (result.to < requestedFrom) {
                     throw new Error(
@@ -333,7 +335,7 @@ class OperationsHistory extends React.Component {
                   }
 
                   const preprocessedOperations = preprocessOperations(
-                    agentConfiguration,
+                    entityConfiguration,
                     result.operations,
                     result.initialState
                   );
@@ -377,12 +379,12 @@ class OperationsHistory extends React.Component {
                 );
               }
 
-              const { agentConfiguration } = this.props;
+              const { entityConfiguration } = this.props;
               const { loadedFrom, loadedOperations, loadedTo } = this.state;
               // Let's deal with 'afterOperations'
               const lastLoadedState = loadedOperations[0].state;
               const preprocessedAfterOperations = preprocessOperations(
-                agentConfiguration,
+                entityConfiguration,
                 afterResults.operations.filter(
                   ({ timestamp }) => timestamp > loadedTo
                 ),
@@ -391,7 +393,7 @@ class OperationsHistory extends React.Component {
 
               // Let's deal with 'beforeOperations' && 'beforeInitialState'
               const preprocessedBeforeOperations = preprocessOperations(
-                agentConfiguration,
+                entityConfiguration,
                 beforeResults.operations.filter(
                   ({ timestamp }) => timestamp < loadedFrom
                 ),
@@ -522,7 +524,7 @@ class OperationsHistory extends React.Component {
     );
   }
 
-  componentDidUpdate(prevProps, prevState) {
+  componentDidUpdate(prevProps) {
     const { scrollToTimestamp } = this.state;
     if (scrollToTimestamp != null) {
       // Desired offset was set to something, that triggered a scroll to the offset
@@ -532,9 +534,9 @@ class OperationsHistory extends React.Component {
       });
     }
     if (this.props.focus != prevProps.focus) {
-      this.setState((state) => ({
+      this.setState({
         scrollToTimestamp: this.props.focus
-      }));
+      });
     }
     if (this.props.initialOperations !== prevProps.initialOperations) {
       // New initial operations, it's like a new start.
@@ -565,7 +567,7 @@ class OperationsHistory extends React.Component {
   }
 
   render() {
-    const { height, rowHeight, width } = this.props;
+    const { entityConfiguration, height, rowHeight, width } = this.props;
     const { error, estimatedCount, scrollToTimestamp } = this.state;
 
     if (error) {
@@ -580,7 +582,7 @@ class OperationsHistory extends React.Component {
         maxWidth={ width }
       >
         <thead ref={ this._setHeaderElement }>
-          <HeaderRow properties={ this._extractedProperties } />
+          <HeaderRow properties={ this._extractedProperties } isGenerator={ entityConfiguration.filter !== undefined } />
         </thead>
         <InfiniteList
           tag='tbody'
@@ -608,16 +610,18 @@ OperationsHistory.defaultProps = {
   ) => Promise.reject(new Error('\'onRequestOperations\' is not defined.')),
   rowHeight: 45,
   height: 600,
-  initialOperations: []
+  initialOperations: [],
+  initialState: {}
 };
 
 OperationsHistory.propTypes = {
-  agentConfiguration: PropTypes.object.isRequired,
+  entityConfiguration: PropTypes.object.isRequired,
   onRequestOperations: PropTypes.func,
   rowHeight: PropTypes.number,
   height: PropTypes.number,
   width: PropTypes.number,
   initialOperations: PropTypes.array,
+  initialState: PropTypes.object,
   to: PropTypes.number,
   from: PropTypes.number,
   focus: PropTypes.number